requirements.txt
fastapi
uvicorn
tensorflow
numpy
pandas
scikit-learn
joblib
matplotlib
openai>=1.0.0
python-dotenv

